updated layout and removed old files
