+++ mrxUnitTest/ast_spec.go
@@ -5,8 +5,6 @@ import (
 	"io"
 
 	"github.com/metarex-media/mrx-tool/klv"
-	mxf2go "github.com/metarex-media/mxf-to-go"
-	. "github.com/onsi/gomega"
 )
 
 /*
@@ -48,155 +46,22 @@ testing header
 
 */
 
-type tester struct {
-	header []func(doc io.ReadSeeker, node *MXFNode, tc *TestContext)
-}
-
-type ISXD struct {
-}
-
-func (i ISXD) DocName() string {
-	return "RDD-47:2018"
-}
-
 func genSpec(docName, sections, command string, commandPosition int) string {
 	// is there a parent required
 	return fmt.Sprintf("%s%s%s%v", docName, sections, command, commandPosition)
 }
-func (i ISXD) TestHeader(doc io.ReadSeeker, header *PartitionNode) func(t Test) {
 
-	return func(t Test) {
-		var isxdDesc *Node
-		for _, child := range header.HeaderMetadata {
-			isxdDesc = child.FindSymbol("060e2b34.02530105.0e090502.00000000")
-			if isxdDesc != nil {
-				break
-			}
-		}
-		// rdd-47:2009/11.5.3/shall/4
-		t.Test("Checking that the isxd descriptor is present in the header metadata"+genSpec(i.DocName(), "9.2", "shall", 1), func() bool {
-			return t.Expect(isxdDesc).ShallNot(BeNil())
-		})
-		if isxdDesc != nil {
-			// decode the group
-			isxdDecode, err := DecodeGroupNode(doc, isxdDesc, header.Props.Primer)
-			fmt.Println(isxdDecode, err)
-			t.Test("Checking that the data essence coding filed is present in the isxd descriptor"+genSpec(i.DocName(), "9.3", "shall", 1), func() bool {
-				return t.Expect(isxdDecode["DataEssenceCoding"]).Shall(Equal(mxf2go.TAUID{
-					Data1: 101591860,
-					Data2: 1025,
-					Data3: 261,
-					Data4: mxf2go.TUInt8Array8{14, 9, 6, 6, 0, 0, 0, 0},
-				}))
-			})
-		}
-
-		// handle the static track sections of the path
-		GenericCountPositions := make([]int, 0)
-		for i, part := range header.Parent.Partitions {
-			// check the essence in each partitoin?
-			if part.Props.PartitionType == GenericStreamPartition {
-				// extra check is counting the steamIDs
-				GenericCountPositions = append(GenericCountPositions, i)
-
-			}
-		}
-
-		if len(GenericCountPositions) > 0 {
-			// ibly run if there's any generic essence
-			var staticTrack *Node
-
-			for _, child := range header.HeaderMetadata {
-				staticTrack = child.FindSymbol("060e2b34.027f0101.0d010101.01013a00")
-				if staticTrack != nil {
-					break
-				}
-			}
-
-			t.Test("Checking that a static track is present in the header metadata "+genSpec(i.DocName(), "5.4", "shall", 1), func() bool {
-				return t.Expect(staticTrack).ToNot(BeNil())
-			})
-
-			if staticTrack != nil {
-
-				sequence := staticTrack.FindSymbol("060e2b34.027f0101.0d010101.01010f00")
-				t.Test("Checking that the static track points to a sequence"+genSpec(i.DocName(), "5.4", "shall", 2), func() bool {
-					return t.Expect(sequence).ToNot(BeNil())
-				})
-
-				t.Test("Checking that the static track sequence has as many sequence children as partitions"+genSpec(i.DocName(), "5.4", "shall", 2), func() bool {
-					return t.Expect(len(sequence.Children)).Shall(Equal(len(GenericCountPositions)))
-				})
-			}
-		}
-
-	}
-	// test ISXD descriptor
-
-}
-
-func (i ISXD) TestEssence(doc io.ReadSeeker, header *PartitionNode) func(t Test) {
-
-	return func(t Test) {
-		if header.Props.PartitionType == BodyPartition && len(header.Essence) > 0 {
-			allISXD := true
-
-			pattern := []string{}
-			patternTally := true
-			for _, e := range header.Essence {
-				ess := nodeToKLV(doc, e)
-
-				if fullNameMask(ess.Key, 13, 15) != "060e2b34.01020105.0e090502.017f017f" {
-					allISXD = false
-					break
-				}
-				fullKey := fullNameMask(ess.Key)
-				if len(pattern) != 0 {
-					if pattern[0] == fullKey {
-						patternTally = false
-					} else if patternTally {
-						pattern = append(pattern, fullKey)
-					}
-				} else {
-					pattern = append(pattern, fullKey)
-				}
-
-			}
-
-			t.Test("Checking that the only ISXD essence keys are found in body partitions"+genSpec(i.DocName(), "7.5", "shall", 1), func() bool {
-				return t.Expect(allISXD).Shall(BeTrue(), "Other essence keys found")
-			})
-
-			if allISXD {
-
-				breakPoint := 0
-				for i, e := range header.Essence {
-					ess := nodeToKLV(doc, e)
-
-					if fullNameMask(ess.Key) != pattern[i%len(pattern)] {
-						breakPoint = e.Key.Start
-						break
-					}
-
-				}
-
-				t.Test("Checking that the content package order are regular throughout the essence stream"+genSpec(i.DocName(), "7.5", "shall", 1), func() bool {
-					return t.Expect(breakPoint).Shall(Equal(0), fmt.Sprintf("irregular key found at byte offset %v", breakPoint))
-				})
-			}
-		} else if header.Props.PartitionType == GenericStreamPartition {
-			// check it passes 2057 rules
-		}
-	}
-}
-
-// specificions
-
+// Specification are the test functions for testing an MXF file to
+// a specification
 type Specification interface {
+	// Test the header partition for metadata etc
 	TestHeader(doc io.ReadSeeker, header *PartitionNode) func(t Test)
-	TestEssence()
-	TestStructure()
-	TestExtra()
+	// TestEssence for testing the essence within a partition
+	TestEssence(doc io.ReadSeeker, header *PartitionNode) func(t Test)
+	// test the overall structure of the mxf file
+	TestStructure(doc io.ReadSeeker, mxf *MXFNode) func(t Test)
+	// TestExtra is for any test cases not covered by, header, essence or structure tests
+	TestExtra(doc io.ReadSeeker, mxf *MXFNode) func(t Test)
 	// TestMarker() - something the test has to find in order to run
 }
 
@@ -204,7 +69,7 @@ func MRXTest(doc io.ReadSeeker, w io.Writer, specifications ...Specification) er
 
 	klvChan := make(chan *klv.KLV, 1000)
 
-	ast, genErr := MakeAST(doc, w, klvChan, 10)
+	ast, genErr := MakeAST(doc, klvChan, 10)
 
 	if genErr != nil {
 		return genErr
@@ -220,43 +85,62 @@ func MRXTest(doc io.ReadSeeker, w io.Writer, specifications ...Specification) er
 	tc := NewTestContext(w)
 	defer tc.EndTest()
 
+	tc.structureTest(doc, ast, specifications...)
+
 	for _, part := range ast.Partitions {
 
 		// check the essence in each partitoin?
 		switch part.Props.PartitionType {
 		case HeaderPartition, FooterPartition:
-			tc.headerTest(doc, part)
+			tc.headerTest(doc, part, specifications...)
 		case BodyPartition, GenericStreamPartition:
-			tc.essTest(doc, part)
+			tc.essTest(doc, part, specifications...)
 		case RIPPartition:
-
+			// not sure what happens here yet
 		}
 	}
+
+	tc.extraTest(doc, ast, specifications...)
 
 	return nil
 }
 
+func (tc *TestContext) structureTest(doc io.ReadSeeker, mxf *MXFNode, specifications ...Specification) {
+
+	tc.Header("testing mxf file structure", func(t Test) {
+		for _, spec := range specifications {
+			spec.TestStructure(doc, mxf)(t)
+		}
+	})
+}
+
+func (tc *TestContext) extraTest(doc io.ReadSeeker, mxf *MXFNode, specifications ...Specification) {
+
+	tc.Header("testing mxf file structure", func(t Test) {
+		for _, spec := range specifications {
+			spec.TestExtra(doc, mxf)(t)
+		}
+	})
+}
+
 func (tc *TestContext) headerTest(doc io.ReadSeeker, header *PartitionNode, specifications ...Specification) {
 
-	tc.Header(fmt.Sprintf("testing %s partition at offset %v", header.Props.PartitionType, header.Key.Start), func(t Test) {
+	tc.Header(fmt.Sprintf("testing header %s partition at offset %v", header.Props.PartitionType, header.Key.Start), func(t Test) {
 
 		for _, spec := range specifications {
-			fmt.Println(t, "T HERE")
+
 			spec.TestHeader(doc, header)(t)
 		}
 
 	})
 }
 
-func (tc *TestContext) essTest(doc io.ReadSeeker, header *PartitionNode) {
+func (tc *TestContext) essTest(doc io.ReadSeeker, header *PartitionNode, specifications ...Specification) {
 
-	specifications := []func(doc io.ReadSeeker, header *PartitionNode) func(t Test){ISXD{}.TestEssence}
-
-	tc.Header(fmt.Sprintf("testing %s partition at offset %v", header.Props.PartitionType, header.Key.Start), func(t Test) {
+	tc.Header(fmt.Sprintf("testing essence %s partition at offset %v", header.Props.PartitionType, header.Key.Start), func(t Test) {
 
 		for _, spec := range specifications {
-			fmt.Println(t, "T HERE")
-			spec(doc, header)(t)
+			spec.TestEssence(doc, header)(t)
 		}
 
 	})
