+++ mrxUnitTest/tester1.log
@@ -1,86 +1,53 @@
-Running Partition 0 MetaData Tests tests:
-	Checking the this partition pointer matches the actual byte offset of the file
-        Pass
+Running testing mxf file structure tests:
+	    Checking that the generic partition positions match the expected positions at the end of the fileRDD-47:20185.4shall3
+            Pass
 Ran 1 tests: Passed:1 , Failed: 0
-Running Partition 0 Tests tests:
-	Checking the this partition pointer matches the actual byte offset of the file
-        Pass
-	Checking the previous partition pointer is the correct byte position
-        Pass
-	Checking the header metadata count matches the actual count of the metadata
-        Pass
-	Checking if the partition contains the group 060e2b34.027f0101.0d010201.01050100
-        Pass
-Ran 4 tests: Passed:4 , Failed: 0
-Running Partition 1 MetaData Tests tests:
-	Checking the this partition pointer matches the actual byte offset of the file
-        Pass
-Ran 1 tests: Passed:1 , Failed: 0
-Running Partition 1 Tests tests:
-	Checking the this partition pointer matches the actual byte offset of the file
-        Pass
-	Checking the previous partition pointer is the correct byte position
-        Pass
-	Checking the header metadata count matches the actual count of the metadata
-        Pass
-	Checking if the partition contains the group 060e2b34.027f0101.0d010201.01050100
-        Fail!
-            the group was not found
-            Expected
-            	[]byte(nil)
-            to contain the group ID of
-            	"060e2b34.027f0101.0d010201.01050100"
-Ran 4 tests: Passed:3 , Failed: 1
-Running Partition 1 Essence Tests tests:
-	Checking the essence keys do not change order
-        Pass
-	Checking the metarex essence keys are located in the correct partition types
-        Pass
-	Checking the metarex essence keys have the correct element number and count
-        Pass
-Ran 3 tests: Passed:3 , Failed: 0
-Running Partition 2 MetaData Tests tests:
-	Checking the this partition pointer matches the actual byte offset of the file
-        Pass
-Ran 1 tests: Passed:1 , Failed: 0
-Running Partition 2 Tests tests:
-	Checking the this partition pointer matches the actual byte offset of the file
-        Pass
-	Checking the previous partition pointer is the correct byte position
-        Pass
-	Checking the header metadata count matches the actual count of the metadata
-        Pass
-	Checking if the partition contains the group 060e2b34.027f0101.0d010201.01050100
-        Fail!
-            the group was not found
-            Expected
-            	[]byte(nil)
-            to contain the group ID of
-            	"060e2b34.027f0101.0d010201.01050100"
-Ran 4 tests: Passed:3 , Failed: 1
-Running Partition 2 Essence Tests tests:
-	Checking the essence keys do not change order
-        Pass
-	Checking the metarex essence keys are located in the correct partition types
-        Pass
-	Checking the metarex essence keys have the correct element number and count
-        Pass
-Ran 3 tests: Passed:3 , Failed: 0
-Running Partition 3 MetaData Tests tests:
-	Checking the this partition pointer matches the actual byte offset of the file
-        Pass
-Ran 1 tests: Passed:1 , Failed: 0
-Running Partition 3 Tests tests:
-	Checking the this partition pointer matches the actual byte offset of the file
-        Pass
-	Checking the previous partition pointer is the correct byte position
-        Pass
-	Checking the header metadata count matches the actual count of the metadata
-        Pass
-	Checking if the partition contains the group 060e2b34.027f0101.0d010201.01050100
-        Pass
-Ran 4 tests: Passed:4 , Failed: 0
-Running Random Index Pack Tests tests:
-	Checking the partition positions in the file match those in the supplied random index pack
-        Pass
-Ran 1 tests: Passed:1 , Failed: 0
+Running testing header header partition at offset 0 tests:
+	    Checking that the isxd descriptor is present in the header metadataRDD-47:20189.2shall1
+            Pass
+	    Checking that the data essence coding filed is present in the isxd descriptorRDD-47:20189.3shall1
+            Pass
+	    Checking that a static track is present in the header metadata RDD-47:20185.4shall1
+            Pass
+	    Checking that the static track points to a sequenceRDD-47:20185.4shall2
+            Pass
+	    Checking that the static track sequence has as many sequence children as partitionsRDD-47:20185.4shall2
+            Fail!
+                Expected
+                    <int>: 2
+                to equal
+                    <int>: 3
+Ran 5 tests: Passed:4 , Failed: 1
+Running testing essence body partition at offset 3510 tests:
+	    Checking that the only ISXD essence keys are found in body partitionsRDD-47:20187.5shall1
+            Fail!
+                Other essence keys found
+                Expected
+                    <bool>: false
+                to be true
+Ran 1 tests: Passed:0 , Failed: 1
+Running testing essence generic stream partition partition at offset 4046 tests:
+Ran 0 tests: Passed:0 , Failed: 0
+Running testing essence generic stream partition partition at offset 4226 tests:
+Ran 0 tests: Passed:0 , Failed: 0
+Running testing essence generic stream partition partition at offset 4404 tests:
+Ran 0 tests: Passed:0 , Failed: 0
+Running testing header footer partition at offset 6756 tests:
+	    Checking that the isxd descriptor is present in the header metadataRDD-47:20189.2shall1
+            Pass
+	    Checking that the data essence coding filed is present in the isxd descriptorRDD-47:20189.3shall1
+            Pass
+	    Checking that a static track is present in the header metadata RDD-47:20185.4shall1
+            Pass
+	    Checking that the static track points to a sequenceRDD-47:20185.4shall2
+            Pass
+	    Checking that the static track sequence has as many sequence children as partitionsRDD-47:20185.4shall2
+            Fail!
+                Expected
+                    <int>: 2
+                to equal
+                    <int>: 3
+Ran 5 tests: Passed:4 , Failed: 1
+Running testing mxf file structure tests:
+Ran 0 tests: Passed:0 , Failed: 0
+Test Failed
