+++ mrxUnitTest/tester2.log
@@ -1,0 +1,97 @@
+Running testing mxf file structure tests:
+	    Checking that the generic partition positions match the expected positions at the end of the fileRDD-47:20185.4shall3
+            Pass
+Ran 1 tests: Passed:1 , Failed: 0
+Running testing header header partition at offset 0 tests:
+	    Checking that the isxd descriptor is present in the header metadataRDD-47:20189.2shall1
+            Pass
+	    Checking that the data essence coding filed is present in the isxd descriptorRDD-47:20189.3shall1
+            Pass
+	    Checking that a static track is present in the header metadata RDD-47:20185.4shall1
+            Pass
+	    Checking that the static track points to a sequenceRDD-47:20185.4shall2
+            Pass
+	    Checking that the static track sequence has as many sequence children as partitionsRDD-47:20185.4shall2
+            Pass
+Ran 5 tests: Passed:5 , Failed: 0
+Running testing essence body partition at offset 16384 tests:
+	    Checking that the only ISXD essence keys are found in body partitionsRDD-47:20187.5shall1
+            Pass
+	    Checking that the content package order are regular throughout the essence streamRDD-47:20187.5shall1
+            Pass
+Ran 2 tests: Passed:2 , Failed: 0
+Running testing essence body partition at offset 412986 tests:
+Ran 0 tests: Passed:0 , Failed: 0
+Running testing essence body partition at offset 415906 tests:
+	    Checking that the only ISXD essence keys are found in body partitionsRDD-47:20187.5shall1
+            Pass
+	    Checking that the content package order are regular throughout the essence streamRDD-47:20187.5shall1
+            Pass
+Ran 2 tests: Passed:2 , Failed: 0
+Running testing essence body partition at offset 814024 tests:
+Ran 0 tests: Passed:0 , Failed: 0
+Running testing essence body partition at offset 816944 tests:
+	    Checking that the only ISXD essence keys are found in body partitionsRDD-47:20187.5shall1
+            Pass
+	    Checking that the content package order are regular throughout the essence streamRDD-47:20187.5shall1
+            Pass
+Ran 2 tests: Passed:2 , Failed: 0
+Running testing essence body partition at offset 1214633 tests:
+Ran 0 tests: Passed:0 , Failed: 0
+Running testing essence body partition at offset 1217553 tests:
+	    Checking that the only ISXD essence keys are found in body partitionsRDD-47:20187.5shall1
+            Pass
+	    Checking that the content package order are regular throughout the essence streamRDD-47:20187.5shall1
+            Pass
+Ran 2 tests: Passed:2 , Failed: 0
+Running testing essence body partition at offset 1615042 tests:
+Ran 0 tests: Passed:0 , Failed: 0
+Running testing essence body partition at offset 1617962 tests:
+	    Checking that the only ISXD essence keys are found in body partitionsRDD-47:20187.5shall1
+            Pass
+	    Checking that the content package order are regular throughout the essence streamRDD-47:20187.5shall1
+            Pass
+Ran 2 tests: Passed:2 , Failed: 0
+Running testing essence body partition at offset 2015753 tests:
+Ran 0 tests: Passed:0 , Failed: 0
+Running testing essence body partition at offset 2018673 tests:
+	    Checking that the only ISXD essence keys are found in body partitionsRDD-47:20187.5shall1
+            Pass
+	    Checking that the content package order are regular throughout the essence streamRDD-47:20187.5shall1
+            Pass
+Ran 2 tests: Passed:2 , Failed: 0
+Running testing essence body partition at offset 2416773 tests:
+Ran 0 tests: Passed:0 , Failed: 0
+Running testing essence body partition at offset 2419693 tests:
+	    Checking that the only ISXD essence keys are found in body partitionsRDD-47:20187.5shall1
+            Pass
+	    Checking that the content package order are regular throughout the essence streamRDD-47:20187.5shall1
+            Pass
+Ran 2 tests: Passed:2 , Failed: 0
+Running testing essence body partition at offset 2820711 tests:
+Ran 0 tests: Passed:0 , Failed: 0
+Running testing essence body partition at offset 2823631 tests:
+	    Checking that the only ISXD essence keys are found in body partitionsRDD-47:20187.5shall1
+            Pass
+	    Checking that the content package order are regular throughout the essence streamRDD-47:20187.5shall1
+            Pass
+Ran 2 tests: Passed:2 , Failed: 0
+Running testing essence body partition at offset 3177314 tests:
+Ran 0 tests: Passed:0 , Failed: 0
+Running testing essence generic stream partition partition at offset 3179926 tests:
+Ran 0 tests: Passed:0 , Failed: 0
+Running testing header footer partition at offset 3183692 tests:
+	    Checking that the isxd descriptor is present in the header metadataRDD-47:20189.2shall1
+            Fail!
+                Expected
+                    <*mrxUnitTest.Node | 0x0>: nil
+                not to be nil
+	    Checking that a static track is present in the header metadata RDD-47:20185.4shall1
+            Fail!
+                Expected
+                    <*mrxUnitTest.Node | 0x0>: nil
+                not to be nil
+Ran 2 tests: Passed:0 , Failed: 2
+Running testing mxf file structure tests:
+Ran 0 tests: Passed:0 , Failed: 0
+Test Failed
